Moved legacy bitmask functions to legacy_bitmask.cpp/hpp. Updated includes.
