+++ cpp/src/bitmask/legacy_bitmask.cpp
@@ -14,9 +14,38 @@
  * limitations under the License.
  */
 
+#include "legacy_bitmask.hpp"
 #include <cudf.h>
 #include <utilities/cudf_utils.h>
 
-gdf_size_type gdf_valid_allocation_size(gdf_size_type column_size){
-    return gdf_get_num_bytes_for_valids_allocation(column_size);
+namespace {
+
+// Buffers are padded to 64-byte boundaries (for SIMD) static
+static constexpr int32_t kArrowAlignment = 64;
+
+// Tensors are padded to 64-byte boundaries static
+static constexpr int32_t kTensorAlignment = 64;
+
+// Align on 8-byte boundaries in IPC static
+static constexpr int32_t kArrowIpcAlignment = 8;
+
+// Align on 4-byte boundaries in CUDF static
+static constexpr int32_t kCudfIpcAlignment = 4;
+
+// todo, enable arrow ipc utils, and remove this method
+static gdf_size_type PaddedLength(int64_t nbytes,
+                                  int32_t alignment = kArrowAlignment) {
+  return ((nbytes + alignment - 1) / alignment) * alignment;
+}
+
+}  // namespace
+
+// Calculates number of bytes for valid bitmask for a column of a specified
+// size
+gdf_size_type gdf_valid_allocation_size(gdf_size_type column_size) {
+  static_assert(sizeof(gdf_valid_type) == 1,
+                "gdf_valid_type assumed to be 1 byte");
+  return PaddedLength(
+      (column_size + (GDF_VALID_BITSIZE - 1)) / GDF_VALID_BITSIZE,
+      kArrowAlignment);
 }
