+++ cpp/src/bitmask/valid_ops.cu
@@ -27,6 +27,7 @@
 #include "rmm/thrust_rmm_allocator.h"
 #include "utilities/error_utils.hpp"
 #include "utilities/cudf_utils.h"
+#include "bitmask/legacy_bitmask.hpp"
 
 #include <thrust/tabulate.h>
 
@@ -140,10 +141,10 @@ gdf_error gdf_count_nonzero_mask(gdf_valid_type const *masks,
 
   // Masks will be proccessed as 4B types, therefore we require that the underlying
   // type be less than or equal to 4B
-  assert(sizeof(valid32_t) >= sizeof(gdf_valid_type));
+  static_assert(sizeof(valid32_t) >= sizeof(gdf_valid_type), "gdf_valid_type is assumed to be <= 4B type");
 
   // Number of gdf_valid_types in the validity bitmask
-  gdf_size_type const num_masks{gdf_get_num_chars_bitmask(num_rows)};
+  gdf_size_type const num_masks{gdf_last_bitmask_index(num_rows)};
 
   // Number of 4 byte types in the validity bit mask 
   gdf_size_type num_masks32{static_cast<gdf_size_type>(std::ceil(static_cast<float>(num_masks) / RATIO))};
@@ -245,7 +246,7 @@ gdf_error gdf_mask_concat(gdf_valid_type *output_mask,
     // as input
     thrust::tabulate(rmm::exec_policy()->on(0),
                      output_mask,
-                     output_mask + gdf_get_num_chars_bitmask(output_column_length),
+                     output_mask + gdf_last_bitmask_index(output_column_length),
                      mask_concatenator);
 
     CUDA_TRY( cudaGetLastError() );
