+++ cpp/src/copying/gather.cu
@@ -22,6 +22,7 @@
 #include "rmm/thrust_rmm_allocator.h"
 #include "utilities/cudf_utils.h"
 #include "utilities/type_dispatcher.hpp"
+#include <bitmask/legacy_bitmask.hpp>
 
 /**
  * @brief Operations for copying from one column to another
