+++ cpp/src/copying/scatter.cu
@@ -20,6 +20,7 @@
 #include "rmm/thrust_rmm_allocator.h"
 #include "utilities/cudf_utils.h"
 #include "utilities/type_dispatcher.hpp"
+#include <bitmask/legacy_bitmask.hpp>
 
 namespace cudf {
