+++ cpp/src/reductions/reductions.cu
@@ -2,6 +2,7 @@
 #include "utilities/cudf_utils.h"
 #include "utilities/error_utils.hpp"
 #include "utilities/type_dispatcher.hpp"
+#include "bitmask/legacy_bitmask.hpp"
 
 #include <cub/block/block_reduce.cuh>
