+++ cpp/src/replace/replace.cu
@@ -24,6 +24,7 @@
 #include "utilities/error_utils.hpp"
 #include "utilities//type_dispatcher.hpp"
 #include "utilities/cudf_utils.h"
+#include "bitmask/legacy_bitmask.hpp"
 
 namespace{ //anonymous
