+++ cpp/tests/groupby/groupby_test.cu
@@ -35,6 +35,7 @@
 // See this header for all of the recursive handling of tuples of vectors
 #include "test_parameters.cuh"
 #include "groupby_test_helpers.cuh"
+#include <bitmask/legacy_bitmask.hpp>
 
 // A new instance of this class will be created for each *TEST(GroupTest, ...)
 // Put all repeated setup and validation stuff here
